refactor: split apart file
