+++ dev-client/src/screens/UserSettingsScreen/components/menu/TosItem.tsx
@@ -1,0 +1,24 @@
+/*
+ * Copyright © 2025 Technology Matters
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see https://www.gnu.org/licenses/.
+ */
+
+import {LinkItem} from 'terraso-mobile-client/screens/UserSettingsScreen/components/menu/LinkItem';
+
+export const TosItem = () => {
+  return (
+    <LinkItem labelI18nKey="general.tos.label" urlI18nKey="general.tos.url" />
+  );
+};
